Add a test program which plays a synthesized tone
Currently 'make check' just checks the API using silence. For
functional testing, it's nice to generate audible output to
verify the code is talking to the platform and hardware correctly.

This commit adds a 'test_tone.c' based on the example code
in the api documentation. It plays a simple synthesized
tone for one second.

Since playing audio automatically can be surprising, I didn't
add it to the tests. It must be run manually.