Import a modified version of an Android header
The SLES/OpenSLES_AndroidConfiguration.h header in the Android NDK has a bug
where the whole file is wrapped in |#ifdef __cplusplus| macro, so we can't get
Android specific OpenSLES values. This patch imports the header, with those
macros removed.

Additionaly, this patch moves the Android-specific headers in an src/android/
directory.