Add a resampler that can synchronously resample audio.
Depending on if its used for input or output audio, it can be used in two
different ways, so it has to expose its input buffer to minimise copy (when the
input of the resampler is output audio, directly written inside an audio
callback), and to be able to expose an internal buffer so that the output of
the resampling process can directly be passed to a callback.

When created, additional latency can be added to the pipeline so that multiple
resamplers of different filter length (=latency) can be synchronized.