Make the resampler handle two opposite streams.
A duplex resampler takes two audio streams (one input and output), with
potentially different sample-rate, and a target rate. It calls back
using the data callback, deliverying input frames at the target rate,
along with a buffer that can be filled with the same number of output
frames at target_rate.

When resampling only one side, a delay line is inserted in the opposite
direction to keep as much synchronization as possible, sacrificing a few
frames of latency. This can help acoustic echo cancelation algorithms.
This is to replace a seemingly no-op resampler, that would end up being
more expensive.

The input rate is the rate of the platform level input stream, i.e. the
sample-rate at which the audio is delivered to cubeb. The output rate is
the rate of the output stream, i.e. the rate at which cubeb should
deliver the audio to the platform API. The target rate is the rate that
was asked for by the user when calling `cubeb_stream_init`.

Consequently, the audio flows like this, where ir and or are input and
output rate, and tr is the target rate:

Platform API --- Input Audio Buffer ---> Duplex Resampler ----.
                         ir              input resampling     |
                                             ir -> tr         |

                                                    prepare output audio
                                                        buffer at tr
                                                              |
                                                    cubeb_data_callback
                                                      input and output
                                                       buffers at tr
                                                              |
                                                              |
Platform API <--- Output Audio Buffer --- Duplex Resampler <--`
                         or              output resampling
                                              tr -> or

The configuration of the processing pipeline is done using
template-based static polymorphism and function pointers.