Add a test for duplex streams.
This simply checks we have (non-silent) input data with a duplex stream,
and copies it to the output, up-mixing it to stereo.

This allows the duplex code to be exercised with a real callback, but
could be more thorough. `test_resampler.cpp` unit tests the resamplers
so we should be somewhat safe nonetheless.