Expose the new duplex resampler in the C API
This also adds a member to the C API, to know the resampler latency. For
now, this is mainly used for testing.

The C API is using `cubeb_params.format` to know the type of the
buffers, and uses `void*` pointer. Right after the C layer, everything
is casted properly to the right types, and templates are used to make
sure we're using the right functions.