Rework how draining happens.
It's now much closer than what was happening before, that was more correct.
When getting back less frames than expected from the callback, we don't
arbitrarily put silent frames at the end, and prefer to track the number of
frames written in the buffer (like WASAPI does with its buffers, in a way).