alsa: Work around audio playback issues with ALSA/dmix configurations.
In some situations (e.g. playing out 44.1kHz audio with a 48kHz dmix
configuration and letting ALSA/dmix resample), we receive wakeups from
poll() that are not aligned with complete refill periods and thus aren't
signalled via POLLOUT.

Rather than relying on POLLOUT, query the available write space and
refill what we can on each wakeup.  This will cause libcubeb's
data_callback to be called somewhat more frequently in these situations,
but it's not clear we have a better option (or how much that will
matter).