Extract audiounit system calls from `audiounit_stream_init`.
This copies a number of piece of state on `cubeb_stream` so that we can keep
them around.
Then, all the system calls from `audiounit_strea_init` have been extracted into
`setup_audiounit_stream`.

The idea is that we can now reset the internal of the stream without having to
`cubeb_stream_destroy`.

This also introduces `has_input` and `has_output` as convenience functions.