Use atomics to detect device switch and not confuse the resampler
This is slightly complex:
- If we're running input and output on two different devices, then multiple
threads are used to service the input and output callbaks, so we need to use
lock-free mechanism.
- We need to keep providing input to the resampler, so we don't buffer, and we
don't trip up the underrun assert.
- We're keeping track of how many output callback in a row
(`stm->output_callback_in_a_row`, atomic) we've had. If it's
too high, we insert silence. This can happen with buggy devices (presumably), or
when switching to a device that takes some time to warm up (USB headsets for
example take some time to start delivering input data).
- We're also keeping track of whether we're switching or not, and insert silence
if we are, and we don't have enough input. This is `stm->switching`, that is
atomic.

This mechanism allows inserting silence input buffer only when strictly
necessary, keeping the input to output latency constant, and the overall latency
constant as well.