Revert Travis CI to Clang 3.6 to match Gecko requirements.
https://developer.mozilla.org/en-US/docs/Using_CXX_in_Mozilla_code

The change to 3.8 (and libc++) was to work around a bug in the libstdc++
headers provided by g++ 4.8 that fail to compile with Clang in C++14 mode.
Later, we reverted from C++14 to C++11 because C++14 is yet available in
Gecko builds.

Also includes a further attempt to clean up the .travis.yml.