audiounit: Prefer a bare function pointer over std::function. (#271)
This addresses BMO bug #1339537, which aims to ban passing structs containing
alignas members by value.  The analysis for this is tripped up by libc++'s
implementation of std::function.

Since we're the only place in Gecko hitting this, and have no real need for
std::function here, the simplest solution is to switch to a bare function
pointer.