wasapi: Implement S16NE-format streams
Brings the WASAPI backend up to parity with the other backends, making
it easier to write cross-platform code for S16 audio. I tried to base my
approach off of the AudioUnit backend (in fact, `auto_array_wrapper` is
now in cubeb_utils and used by both) and a little bit from the Jack
backend (the `bytes_per_sample` member).