Make cubeb work with high channel count devices (#288)
* Bail if the device has too many channels.

If the device on the host where we're trying to open an audio stream has too
many channels, just bail instead of trashing the memory in the loop just below.

* Allow using audio output devices that have up to 256 channels.

In particular, this allows using sound cards that have more than 8 channels
(7.1), for musical purposes, and allows using the 64 output channels of
SoundFlower, for testing and also advanced musical use-cases.

If we are in a topology that has more than 8 channels, the other channels are
considered unmapped, i.e. it's just a channel without a specific speaker
assignation. This is useful for custom setups, such as wave field synthesis, or
other creative applications.

* Address chunmin and achronop comments.