Add backend specification support to common.h
Set CUBEB_BACKEND env var to the name of valid backend supported by
the current platform to prefer that backend over the default choice.

If CUBEB_BACKEND is not valid, backend selection falls back to the
standard behaviour.

A warning is printed to stderr if the chosen backend id doesn't match
the requested one.