sndio: use samples instead of bytes in latency calculations.
The number of samples "seen" by the data call-back can be obtained by
accumulating values it returns. The number of samples "seen" by the
hardware is obtained with the sndio sio_onmove() interface. The
latency is simply the difference between these two counters.

Note that we used to accumulate bytes consumed by sio_write(), which
is slightly more complicated and gives a slightly delayed
representation of the actual latency.