sndio: add recording and full-duplex modes.
The same device is used in full-duplex mode as defined by the sndio
inteface: play and recording always start simultaneously and are
maintained in sync. The N-th recorded sample is recorded exactly when
the N-th playback sample is played. Both directions use the same clock
source. In case of overrun/underrun the system compensates with the
proper amount of silence in order to keep both directions in sync.