Rework get_input_buffer to better function with pull based streams and remove cruft.
Update get_input_buffer to handle 'pull based' streams. Specifically the
existing WASAPI backend expects streams to be event driven: an event will be
fired when input is available and it can then be processed. Provided an event
is fired per packet there is no issue.

However, if less than one event is fired per input, or if the input is buffered
and then 'pulled' then the current impl may fail to process inputs, only
handling the first packet in the buffer.

The changes here attempt to handle this. They also remove some code which
appears to be unneeded. Specifically some of the checks that use
GetCurrentPadding and GetNextPacketSize are not required. These two functions
should return equivalent values for IAudioCaptureClient, and these changes
reflect that. For more info on this please see
https://msdn.microsoft.com/en-us/library/windows/desktop/dd370868(v=vs.85).aspx