add support for recording in sndio backend (#376)
* sndio: fix debug printf() format bug; uint64_t must use PRId64.

* sndio: use samples instead of bytes in latency calculations.

The number of samples "seen" by the data call-back can be obtained by
accumulating values it returns. The number of samples "seen" by the
hardware is obtained with the sndio sio_onmove() interface. The
latency is simply the difference between these two counters.

Note that we used to accumulate bytes consumed by sio_write(), which
is slightly more complicated and gives a slightly delayed
representation of the actual latency.

* sndio: make 48kH the prefered sample rate.

There's no device-independent prefered rate: any rate will work if
sndiod is running, which is the default on any OpenBSD system. If it
isn't, 48kHz is what is most likely to work as most (but not all)
devices support it.

* sndio: modify comment: there's no device-independent minimum latency.

* sndio: implement method stubs to enumerate devices.

Device enumeration is used by certain tests, so they are need to
benefit from the tests. The provided implementation returns only the
default device which is consistent with the rest of implementation of
the sndio backend.

* sndio: prefix playback-related variable names with 'p'.

No object change. The 'p' prefix will help using similar names for the
recording direction, prefixed by 'r'.

* sndio: fix first block size, no behaviour change.

The initial block size was wrong due to confusion between samples and
bytes, but as it's recalculated at each cycle this bug didn't cause
any problems.

* sndio: set state to CUBEB_STATE_ERROR if the device is disconneced.

This may happen when USB audio devices is disconnected or during a
network failure when a remote audio device is being used.

* sndio: properly check for errors using the sio_eof() function.

In non-blocking mode, it's not correct to just check if sio_write()
returns 0, as it's OK for sio_write() to return 0 when the device
blocks.

* sndio: simplify and comment the way we drain playback buffer.

Introduce a simple "eof" flag that's set whenever the data call-back
returns a truncated block. Besides being simpler, this approach will
make easier handling recording.

* sndio: factor error code paths of sndio_stream_init().

* sndio: fix memory leak when stream is destroyed (forgotten free())

* sndio: add recording and full-duplex modes.

The same device is used in full-duplex mode as defined by the sndio
inteface: play and recording always start simultaneously and are
maintained in sync. The N-th recorded sample is recorded exactly when
the N-th playback sample is played. Both directions use the same clock
source. In case of overrun/underrun the system compensates with the
proper amount of silence in order to keep both directions in sync.