Add loopback test file and test for duplex stream with loopback.
Add a test file for loopback tests and add to build script. This file is
currently configured to be Windows only in build scripts, as the WASAPI backend
is currently the only backend supporting the new loopback interface.

Add a test for duplex streams using the default output device and a loopback
from the same device.