WASAPI: Map KSAUDIO_SPEAKER_QUAD and KSAUDIO_SPEAKER_5POINT1 to cubeb layouts.
Update the WASAPI backend to handle two previously unhandled channel layouts.
Before this change cubeb would expect 4 and 5.1 channel layouts to be made up
of front and side left and right channels. Configurations that had back left
and right channels instead of side left and right channels would be set as
undefined layouts. This would lead to certain issues, such as mixing not taking
place.

This change is somewhat of a band-aid in that these channel configurations
could be represented as distinct by introducing further enums into cubeb.h.
However, this provides better support without a large change and modifying the
API. Comments have been inserted into the code to clarify this and link to
tracking information.