pulse: Don't get some default sink info based on the pa_sink_info build-time size
The build-time pa_sink_info might not have the same size as the at run
time. This especially becomes a problem when the struct is larger at
build-time, since we copy its content for the default sink info.

However, we don't need most of the data in there, so create a new struct
holding what we do need, and copy that. This new struct has a constant
size across versions of pulseaudio.

https://bugzilla.mozilla.org/show_bug.cgi?id=1427150