Fix a deadlock
When creating an aggregate device, and a device collection change callback was
installed, a deadlock occured, because the device collection change listener was
being called synchronously.

Use GCD to send a task to a serial queue, and carefuly unlock the mutex in
audiounit_stream_destroy to not deadlock on the queue.