WASAPI: Check that the input side of a stream was successfully setup before pushing silence.
If setup of the input side of a WASAPI stream failed we would still attempt to
push_silence() on the linear_input_buffer. However, if setup fails then
input_buffer_frame_count may not be set correctly. This results in attempting
to push silence to undesirable addresses (0 in my testing).

Fix this by checking the result of input setup before pushing silence.