Non-wasapi backends return errors if loopback is requested.
Loopback is not supported outside of WASAPI, so if it's requested an error is
returned. This error is set to indicate loopback is not supported in most
cases. Though in some cases a generic error is returned to avoid complicating
existing code.

The current API means requesting a loopback for an output stream doesn't make
sense. However, to keep handling simplified both input and output params are
handled in the same or similar fashion by this commit. Both kinds of request
should fail, but no extra indication is given that requesting output loopback
is aberrant (since it will fail anyway).