WASAPI: Hide loopback dummy output from data callback.
Hides the dummy output device, if present, from the data callback. In addition,
ensure the dummy output device has silence pushed to it. This has the benefit
of the loopback recording silence when nothing else is playing, as well as
making sure the dummy device doesn't output an audible signal.