audiounit: Reduce duplicate code in audiounit_stream_get_current_device (#453)
* Add a test for cubeb_stream_get_current_device() and cubeb_stream_device_destroy()

* Reduce duplicates in audiounit_stream_get_current_device()

The audiounit_stream_get_current_device() will set the default device's name for input and output by the following steps:
1. Get default device's data, whose type is uint32
2. Convert the uint32 data into a string
3. Set the string into cubeb_device's name

Hence we can split it into functions to do the above steps:
i. audiounit_get_default_device_data(...)
  Do the step 1
ii. allocate_and_convert_uint32_into_string(...)
  Do the step 2 and 3. It will allocate memory for a string and put the converted string right there. (If we don't allocate and then set the values into the memory at the same time, we need to find a way to check the memory's boundary.)

They are called by a function named audiounit_get_default_device_name(...), with cubeb_device_type parameter to set the device name of input or output. And audiounit_get_default_device_name(...) will be used in audiounit_stream_get_current_device().

* Replace allocate_and_convert_uint32_into_string(...) by convert_uint32_into_string(...)

1. Rename allocate_and_convert_uint32_into_string(...) to convert_uint32_into_string(...)
2. Return a unique_ptr<char[]> from convert_uint32_into_string(...) pointing to an allocated memory

* Hard-coding the convertion from uint32 into string

This change is to make sure the conversion only ever needs to handle 4 bytes.

* Add more information to log when calling audiounit_stream_get_current_device(...)

We add a log with error message when we cannot get datasource data from the devices. However, we don't return an error code in this case since it's quite common when we try getting that data from USB devices. We will convert the datasource data into a string. If there is no data, the string is empty. Instead of logging when we cannot get the datasource data, it's better to log that the converted name is empty. It'll give more meaning (Users are more likely confused about what the empty datasource means.).

* Rename audiounit_get_default_device_data() to audiounit_get_default_device_datasource()

We should use an explicit name for this function since the same device will return different datasources. For example, the default input device on macbook pro will return "imic" if it uses the default internal microphone or "emic" when it uses an external microphone plugged in audio jack.

TODO: it's better to rename audiounit_stream_get_current_device to audiounit_stream_get_current_device_source. The reason is same as above.