Don't use PLII matrices when downmixing.
PLII mix the opposed channel with the reversed phase. It doesn't sound good when used with a common stereo headset; like the left rear channel sounds too much like it's also coming from the right. Additionally, it clips significantly causing unwanted distortion. So we default to a simpler conversion instead: left goes to left only, same for the right channel.

(Part of the fix for BMO 1509875)

Fixes #476