wasapi: fix hns_to_frames being off-by-one
For example, when given a sample rate of 44100 Hz, and a period of 29025
hundred-nanoseconds, hns_to_frames calculates 128.00025 frames, and
rounds up to 129 frames.

The actual result here should be 128, but that can't be encoded directly
as a REFERENCE_TIME; it'd need to be 29024.9433107. I'm guessing WASAPI
just rounds/ceils to the closest whole number.

So, when calculating hns_to_frames, first subtract one from the hns
value to handle these cases.