osx: drain stream properly
The logic for draining a stream was odd. After drain the AU stop was happening on the next callback. That could create race conditions especially when reinit is taking place at the same time. This patch triggers the stop of AU at the end of the same callback and set the shutdown flag in order to exit early any following callback.