Prepare the repo for merging with cubeb central.
On cubeb repoisitory [1] the rust backend (cubeb-pulse-rs) will be moved inside the src/cubeb-pulse-rs folder. In order to manage that conviniently the folder structure of this repository changed and everything is transfered inside src/cubeb-pulse-rs folder. This structure will allow the content of this repository to be moved directly to the appropriate folder of cubeb repository once merge takes place.

[1] https://github.com/kinetiknz/cubeb