pulse: when default device changes update the info (#547)
The problem was that after a default device change the cubeb_get_max_channel_count() returned the wrong number of channels; it returned the device info (including channel count) of the initial default device. In order to update the default device info dynamically, a subscription to the  event is created during context initialization and the device info is updated accordingly.