Only print the filename when logging (#581)
* Only print the filename when logging

This keeps the logging message consistent after
https://github.com/djg/cubeb-rs/pull/50 is merged, which fixes the BMO
162132.

Instead of printing the full path to the file that is currently running,
we should print the filename only. As long as we have the filename and
the line, we know where the log comes from.

* Reformat `LOG_INTERNAL`

Align the `\` used for `LOG_INTERNAL` macro.

* Correct position of `__FILENAME__`

* Define `__FILENAME__` by `__FILE_NAME__` if possible

clang 11 provide a `__FILE_NAME__` [1] macro. We should use it if it
exists.

[1] https://clang.llvm.org/docs/LanguageExtensions.html#builtin-macros

* Use `strrchr` in `string.h` when building by MSVC

There is no `__builtin_strrchr` in MSVC, so we fullback to use the
`strrchr` in `string.h` to trim the path above the current file name.