Don't return early when there is no input data.
This is simply wrong. Input data are accumulated into the buffer, and
the buffer is emptied into the resampler. The resampler controls the
buffering. It's certainly possible to not have input data here, and
having had two packet worth of data in a previous callback.

This fixes constant glitches on a Sony WH1000-mk3, and doesn't seem to
regress the wired/built-in devices I have.