Add a log callback that takes a formatted message.
In some contexts (e.g. cubeb-rs), it is not possible to pass a method
that formats a string (for instance, because `c_variadic` is not stable
in rust, and because it only supports compile-time constant format
strings).

So, add a callback method that takes a `const char*` that has already
been printf-style formatted.