Wait for all frames to be read before requesting stop
With a large AAudio buffer size it became clear that the end of playback
was getting truncated. AAudioStream_requestStop documentation says it
waits for the stream to be drained. The drain does not seem to apply to
all written frames, but rather all read frames for an output stream.

This patch makes us wait for all written frames to be read before
requesting stop.